+++ classes/suggested-tasks/providers/integrations/yoast/class-organization-logo.php
@@ -126,7 +126,7 @@ class Organization_Logo extends Yoast_Provider {
 	public function add_task_actions( $data = [], $actions = [] ) {
 		$actions[] = [
 			'priority' => 10,
-			'html'     => '<a class="prpl-tooltip-action-text" href="' . \admin_url( 'admin.php?page=wpseo_page_settings#/site-representation' ) . '" target="_blank">' . \esc_html__( 'Set logo', 'progress-planner' ) . '</a>',
+			'html'     => '<a class="prpl-tooltip-action-text" href="' . \admin_url( 'admin.php?page=wpseo_page_settings#/site-representation' ) . '" target="_self">' . \esc_html__( 'Set logo', 'progress-planner' ) . '</a>',
 		];
 
 		return $actions;